+++ crates/ruma-common/CHANGELOG.md
@@ -11,6 +11,7 @@ Improvements:
 * Add `MatrixVersion::V1_4`
 * Stabilize default room server ACL push rule
 * Stabilize `room_types` in `directory::Filter` and `room_type` in `directory::PublicRoomsChunk`
+* Stabilize support for private read receipts
 
 # 0.10.3